Initialize the EssentialHeader of ECDSA JWKs.
The jwk.NewEcdsaPublicKey and jwk.NewEcdsaPrivateKey() functions
are initializing the EssentialHeader of the EcdsaPublicKey struct
which means they do not have an initialized KeyType (jwa.EC) and
attempts to set header parameters with Set() cause a panic. This
initialization is done correctly for RSA keys.

This commit fixes the initializion of the EssentialHeader for JWK
ECDSA keys and sets their key type parameter to jwa.EC. A unit test
is also implemented to verify correct behavior.