Fix bad trimming of compact JWS signatures.
The decoding of compact form JWS tokens incorrectly performs a
TrimRight to strip trailing zero bytes from the token signature
after decoding from base64. It should not do this.

I found that about 1% of the JWS tokens we generated using
another tool were failing signature verification by go-jwx.
These JWS tokens were valid and correctly formatted and after
some spelunking into the go-jwx code we discovered this trim
of trailing zeroes from the compact signature was the problem.
This specifically fails when the generated signature (ECDSA
P-384 in this case) actually had a trailing zero byte in the
serialized S component of the signature.

This patch removes the offending TrimRight. A unit test has
also been added to verify one of the specific instances of
a failing JWS token from my testing. This trimming does not
appear to be done for JSON encoded JWS tokens.