Do not allow unregistered urls to be Fetch()ed
In 98f46c62ddac6e279d5325e8c30ae1723bb620ce I spoke too soon.
Turns out that there's a subtle user interface problem with allowing
`Fetch()` to carry options.

For example:

  af.Fetch(url, options...)
  af.Fetch(url)

When called in this order, the second call effectively resets the
entire option settings from the earlier call. This is okay if you
know exactly what's going on, but I must say that it is definitely
extremely hard for the casual users. You wouldn't really be able to
figure out why the refresh settings from the earlier call is completely
changed after the second call.

So while it is definitely inconvenient, we must introduce a separate
step to configure each URL:

  af.Configure(url, options...)
  af.Fetch(url) // no options allowed
  af.Fetch(url) // no way to botch the previous settings

AddJitionally, it might be plausible to have something along the lines
of only-run-configure-if-this-url-has-not-been-registered:

  for {
    // no op if it's already registered
    af.ConfigureIfNotRegistered(url, options)
    af.Fetch(url)
    ...
  }

Then it is safe to call this in a loop. ATM I'm not sure if it's worth
it, though.