Simplify the validator interface
ClaimsValidator is now just Validator.
This is because... well, we don't have multiple validators, so
we're good with just a Validator.

SingleClaimValidator has been removed.
I tried various implementations of a generic "single claim validator"
API, but I just couldn't find a good, generic, and yet expressive
way of doing this.

Whatever way you use, you end up having to write more code or
reinvent "reflect" to be generic enough. I think leaving it open
with jwt.Validator and jwt.ValidatorFunc is the way to go until
there comes somebody with a clever enough implementation if there is
one.