jwt: only infer algorithm on the already-selected key (#523)
* jwt/test: fix JWTParseVerify test to actually iterate over all the cases

Due to how variable capture works, it was only testing the last key and
alg for the given combination. Test them all.

* jwt/test: improve the dummy keys generated to match the originals

Also sets a kid on the dummy key.

* jwt: only infer algorithm on the already-selected key

We've already done the work to figure out which key to use. Keep using
it when inferring which verification algorithm to use.

Closes #522