Allow matching of JWKS keys with not unique kid for signature verification
This changes the matching of JWKS keys with not unique kid for signature verification.
In case a JWKS store contains keys with not unique kid, all matching keys are selected. The selected keys are then tried to see if one produces a correct signature.
Not unique kid are allowed by the standard
https://datatracker.ietf.org/doc/html/rfc7517#section-4.5