Add `b64: false` header to detached example
When using JWS to sign a detached webhook payload I intend to transmit in the clear, I found the receiving service written in NodeJS (using the very popular `jose` library) assumed the payload was base64-encoded. It demanded the `b64: false` header be sent in order to treat it as plaintext.

While RFC 7797 states this header is OPTIONAL, in order to maximize compatibility with other software and libraries, this example should endeavour to use it. I'm going on the assumption that detached payloads are not likely manually base64-encoded